[WXH-35] Config endpoints support different volume types

+++ source/apis/apis.thrift
@@ -9,7 +9,7 @@ enum VolumeType {
 struct VolumeSettings {
        1: required i32 maxObjectSizeInBytes,
        2: required VolumeType volumeType,
-       3: optional i32 blockDeviceSizeInBytes
+       3: optional i64 blockDeviceSizeInBytes
 }
 
 struct VolumeDescriptor {
